+++ src/modules/user/dto/index.ts
@@ -1,0 +1,26 @@
+// This file will contain all dto for easier import
+
+import { LoginResponseDTO, ReturnUserData } from './login-response.dto';
+import { LogoutResponseDTO } from './logout-response.dto';
+import {
+  RegisterResponseDTO,
+  RegisteredUserData,
+} from './register-response.dto';
+import { RegisterUserInput } from './register-user.input';
+import { UpdateUserPersonalInfoInput } from './update-user-personal-info.input';
+import {
+  UpdateUserResponseDTO,
+  UpdatedUserData,
+} from './update-user-response.dto';
+
+export {
+  LoginResponseDTO,
+  ReturnUserData,
+  LogoutResponseDTO,
+  RegisterResponseDTO,
+  RegisteredUserData,
+  RegisterUserInput,
+  UpdateUserPersonalInfoInput,
+  UpdateUserResponseDTO,
+  UpdatedUserData,
+};
